major gradient changes!
1. The bpsf gradient has been double checked and the equations were changed!
There may have been a bug in the angular symmetry functions, where
when the function was decomposed into a(x), b(x), and c(x), we only
summed

da(x)/dx + db(x)/dx + dc(x)/dx * a(x) * b*(x)

This has been rectified. I am not sure if this was the cause of the
strange discrepancies I have been noticing in the gradient lately or
not.

I will add all of the right answers to the tex file shortly.

2. It is possible to reorder the cartesian gradients from an input H5
file in the same order that the basis sets are in order to facilitate
neural network training.

3. Started fixing some of the test files but will complete the rest
during the next commit.